jurisconsultor/requirements.txt
fastapi
uvicorn
sentence-transformers
pymongo
psycopg2-binary
python-dotenv
requests
beautifulsoup4
numpy
torch
pypdf
python-docx
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.0.1
python-multipart
pydantic==2.8.2
pydantic-core==2.20.1

# Testing Dependencies
pytest
pytest-mock

# LangGraph Dependencies
langchain==0.2.11
langgraph==0.2.38
langchain-openai==0.1.15
langchain-community==0.2.0
langgraph-checkpoint-mongodb==0.1.0
APScheduler==3.10.4
tenacity==8.2.3
